requirements.txt
bitarray==1.7.0
Pillow==8.1.0
